+++ src/test/java/com.outsystems.plugins.inappbrowser/osinappbrowserlib/helpers/OSIABCustomTabsSessionHelperMock.kt
@@ -3,16 +3,20 @@ package com.outsystems.plugins.inappbrowser.osinappbrowserlib.helpers
 import android.content.ComponentName
 import android.content.Context
 import androidx.browser.customtabs.CustomTabsSession
+import com.outsystems.plugins.inappbrowser.osinappbrowserlib.OSIABEvents
 
 class OSIABCustomTabsSessionHelperMock: OSIABCustomTabsSessionHelperInterface {
     private val componentName = "OSIABTestComponent"
+    var eventToReturn: OSIABEvents? = null
+
     override suspend fun generateNewCustomTabsSession(
-        context: Context,
-        onEventReceived: (Int) -> Unit,
+        context: Context, 
+        onEventReceived: (OSIABEvents) -> Unit,
         customTabsSessionCallback: (CustomTabsSession?) -> Unit
     ) {
-        customTabsSessionCallback(
-            CustomTabsSession.createMockSessionForTesting(ComponentName(context, componentName))
-        )
+        if (eventToReturn != null) {
+            onEventReceived(eventToReturn!!)
+        }
+        customTabsSessionCallback(CustomTabsSession.createMockSessionForTesting(ComponentName(context, componentName)))
     }
 }