+++ CHANGELOG.md
@@ -4,9 +4,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Features
+- Add permissions requests and opening file chooser to `OpenInWebView` feature (https://outsystemsrd.atlassian.net/browse/RMET-3534).
+- Add error and loading screens for `OpenInWebView` feature (https://outsystemsrd.atlassian.net/browse/RMET-3492).
+- Add browser events to `OpenInSystemBrowser` feature (https://outsystemsrd.atlassian.net/browse/RMET-3431).
 - Add `OpenInSystemBrowser` (https://outsystemsrd.atlassian.net/browse/RMET-3424).
 - Add UI customizations to`OpenInWebView` (https://outsystemsrd.atlassian.net/browse/RMET-3490).
 - Add `OpenInWebView` with current features and default
   UI (https://outsystemsrd.atlassian.net/browse/RMET-3426).
-- Add browser evetns to `OpenInWebView` feature (https://outsystemsrd.atlassian.net/browse/RMET-3432).
+- Add browser events to `OpenInWebView` feature (https://outsystemsrd.atlassian.net/browse/RMET-3432).
 - Add `OpenInExternalBrowser` (https://outsystemsrd.atlassian.net/browse/RMET-3422).